+++ src/main/java/net/luis/utils/util/Range.java
@@ -34,7 +34,7 @@ public class Range {
 	}
 	
 	public static @NotNull Range of(char min, char max) {
-		return of((int) min, (int) max);
+		return of(min, (int) max);
 	}
 	
 	public int getMin() {
